+++ src/main/java/commands/TBHelpCommand.java
@@ -1,7 +1,7 @@
 package commands;
 
 import net.dv8tion.jda.api.EmbedBuilder;
-import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
+import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
 import java.awt.*;
@@ -37,7 +37,8 @@ public class TBHelpCommand extends ListenerAdapter {
      *
      * @param event guild message event
      */
-    public void onGuildMessageReceived(@Nonnull GuildMessageReceivedEvent event) {
+    @Override
+    public void onMessageReceived(@Nonnull MessageReceivedEvent event) {
         // grab message
         String message = event.getMessage().getContentRaw();
 
@@ -56,7 +57,7 @@ public class TBHelpCommand extends ListenerAdapter {
             eb.addField("`Version 1.0`", msg, true);
 
             // reply
-            event.getChannel().sendMessage(eb.build()).queue();
+            event.getChannel().sendMessageEmbeds(eb.build()).queue();
         }
 
     }